tests/requirements-test.txt
pytest==8.3.4
pytest-html==4.1.1
pytest-env==1.1.5
requests==2.32.3
psycopg2-binary==2.9.10
pydantic==2.10.6
python-dotenv==1.0.1
